+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -27,8 +27,7 @@
 * Deprecate `deserialiseRewardAcnt` in favor of `deserialiseRewardAccount`
 * Deprecate `serialiseRewardAcnt` in favor of `serialiseRewardAccount`
 * Deprecate `RewardAcnt` in favor of `RewardAccount`
-* Add `toRationalJSON` and `parseAsRational` JSON Aeson helpers
-* Nodify `Prices` JSON instances to match `cardano-api`
+* Modify `Prices` and `Nonce` JSON instances to match `cardano-api`
 
 ### `testlib`